# contract/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "ConnectRedemption": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "type": "string"
          },
          "deviceName": {
            "description": "Shown in the account's devices list, e.g. \"Living room Apple TV\".",
            "maxLength": 60,
            "minLength": 1,
            "type": "string"
          },
          "platform": {
            "enum": [
              "ios",
              "ipados",
              "tvos",
              "android",
              "androidtv"
            ],
            "type": "string"
          }
        },
        "required": [
          "code",
          "deviceName",
          "platform"
        ],
        "type": "object"
      },
...
      "DeviceSignIn": {
        "additionalProperties": false,
        "properties": {
          "deviceName": {
            "description": "Shown in the account's devices list, e.g. \"Living room Apple TV\".",
            "maxLength": 60,
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          },
          "platform": {
            "enum": [
              "ios",
              "ipados",
              "tvos",
              "android",
              "androidtv"
            ],
            "type": "string"
          },
          "username": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "deviceName",
          "platform"
        ],
        "type": "object"
      },
      "DeviceToken": {
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "format": "uuid",
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "token",
          "deviceId",
          "user"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "PairingPoll": {
        "additionalProperties": false,
        "properties": {
          "deviceCode": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "deviceCode"
        ],
        "type": "object"
      },
...
      "PairingStatus": {
        "additionalProperties": false,
        "properties": {
          "device": {
            "$ref": "#/components/schemas/DeviceToken"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "denied",
              "expired"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
      "ServerInfo": {
        "additionalProperties": false,
        "properties": {
          "accent": {
            "description": "The site accent colour as a CSS hex value.",
            "type": "string"
          },
          "apiLevel": {
            "description": "Raised whenever clients need new server behaviour.",
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "description": "Stable for the server's lifetime; clients key their accounts by it.",
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "description": "The release the server was built from.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "apiLevel",
          "accent"
        ],
        "type": "object"
      },
...
      "User": {
        "additionalProperties": false,
        "properties": {
          "avatarId": {
            "description": "Artwork id of the profile picture.",
            "type": [
              "string",
              "null"
            ]
          },
          "bannerId": {
            "description": "Artwork id of the profile banner.",
            "type": [
              "string",
              "null"
            ]
          },
          "bio": {
            "description": "Markdown, rendered with raw HTML disabled.",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "role": {
            "enum": [
              "admin",
              "member"
            ],
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "displayName",
          "role",
          "disabled",
          "avatarId",
          "bannerId",
          "bio",
          "createdAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerDevice": {
        "description": "Native clients: a device token from signInDevice, pollPairing or connectDevice.",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieSession": {
        "description": "Browsers.",
        "in": "cookie",
        "name": "couchverse_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Couchverse",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/connect": {
      "post": {
        "operationId": "connectDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectRedemption"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceToken"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/pairings/poll": {
      "post": {
        "operationId": "pollPairing",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingPoll"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingStatus"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/token": {
      "post": {
        "operationId": "signInDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceSignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceToken"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/server": {
      "get": {
        "operationId": "getServer",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerInfo"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "system"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}